+++ src/main/java/ivanwidjanarko/jwork/DatabaseRecruiterPostgre.java
@@ -12,7 +12,7 @@ import static ivanwidjanarko.jwork.DatabaseConnectionPostgre.connection;
  * Class for Database Recruiter Postgre
  *
  * @author Ivan Widjanarko
- * @version 19-06-2021
+ * @version 25-06-2021
  */
 public class DatabaseRecruiterPostgre {
 
@@ -51,7 +51,6 @@ public class DatabaseRecruiterPostgre {
         catch (SQLException e) {
             e.printStackTrace();
             System.err.println(e.getClass().getName() + ":" + e.getMessage());
-            System.exit(0);
         }
         return recruiterObject;
     }
@@ -77,7 +76,6 @@ public class DatabaseRecruiterPostgre {
         catch (SQLException e) {
             e.printStackTrace();
             System.err.println(e.getClass().getName() + ":" + e.getMessage());
-            System.exit(0);
         }
         return value;
     }
@@ -86,23 +84,36 @@ public class DatabaseRecruiterPostgre {
      * Method for get Recruiter by its ID
      * @param recruiter_id Recruiter's ID
      * @return    recruiter
+     * @throws RecruiterNotFoundException Exception if Recruiter not found
      */
-    public static Recruiter getRecruiterById(int recruiter_id) {
+    public static Recruiter getRecruiterById(int recruiter_id) throws RecruiterNotFoundException {
         Recruiter value = null;
         c = connection();
         try {
             stmt = c.createStatement();
-            ResultSet rs = stmt.executeQuery("SELECT * FROM recruiter_location WHERE recruiter_id = "+ recruiter_id + ";");
+            ResultSet rs = stmt.executeQuery("SELECT count(*) FROM" +
+                    "(SELECT * FROM recruiter WHERE id = " + recruiter_id +
+                    ") AS id;");
             while (rs.next()) {
-                recruiter_id = rs.getInt("recruiter_id");
-                String recruiter_name = rs.getString("recruiter_name");
-                String recruiter_email = rs.getString("recruiter_email");
-                String recruiter_phoneNumber = rs.getString("recruiter_phone_number");
-                String location_province = rs.getString("location_province");
-                String location_city = rs.getString("location_city");
-                String location_description = rs.getString("location_description");
-                value = new Recruiter(recruiter_id, recruiter_name, recruiter_email, recruiter_phoneNumber,
-                        new Location(location_province, location_city, location_description));
+                int count = rs.getInt("count");
+
+                if (count == 0) {
+                    throw new RecruiterNotFoundException(recruiter_id);
+                }
+                else {
+                    rs = stmt.executeQuery("SELECT * FROM recruiter_location WHERE recruiter_id = " + recruiter_id + ";");
+                    while (rs.next()) {
+                        recruiter_id = rs.getInt("recruiter_id");
+                        String recruiter_name = rs.getString("recruiter_name");
+                        String recruiter_email = rs.getString("recruiter_email");
+                        String recruiter_phoneNumber = rs.getString("recruiter_phone_number");
+                        String location_province = rs.getString("location_province");
+                        String location_city = rs.getString("location_city");
+                        String location_description = rs.getString("location_description");
+                        value = new Recruiter(recruiter_id, recruiter_name, recruiter_email, recruiter_phoneNumber,
+                                new Location(location_province, location_city, location_description));
+                    }
+                }
             }
             stmt.close();
             c.close();
@@ -111,7 +122,6 @@ public class DatabaseRecruiterPostgre {
         catch (SQLException e) {
             e.printStackTrace();
             System.err.println(e.getClass().getName() + ":" + e.getMessage());
-            System.exit(0);
         }
         return value;
     }
@@ -149,7 +159,6 @@ public class DatabaseRecruiterPostgre {
         catch (SQLException e) {
             e.printStackTrace();
             System.err.println(e.getClass().getName() + ":" + e.getMessage());
-            System.exit(0);
             return null;
         }
         return recruiter;
@@ -159,21 +168,33 @@ public class DatabaseRecruiterPostgre {
      * Method for remove the recruiter from database
      * @param id recruiter's ID
      * @return    recruiter deletion status
+     * @throws RecruiterNotFoundException Exception if Recruiter not found
      */
-    public static boolean removeRecruiter(int id) {
+    public static boolean removeRecruiter(int id) throws RecruiterNotFoundException {
         c = connection();
         try {
             stmt = c.createStatement();
-            String query = "DELETE FROM location WHERE id = "+ id + ";";
-            stmt.executeUpdate(query);
-            stmt.close();
-            c.close();
+            ResultSet rs = stmt.executeQuery("SELECT count(*) FROM" +
+                    "(SELECT * FROM recruiter WHERE id = " + id +
+                    ") AS id;");
+            while (rs.next()) {
+                int count = rs.getInt("count");
+
+                if (count == 0) {
+                    throw new RecruiterNotFoundException(id);
+                }
+                else {
+                    String query = "DELETE FROM location WHERE id = " + id + ";";
+                    stmt.executeUpdate(query);
+                    stmt.close();
+                    c.close();
+                }
+            }
         }
 
         catch (Exception e) {
             e.printStackTrace();
             System.err.println(e.getClass().getName() + ":" + e.getMessage());
-            System.exit(0);
         }
         return true;
     }
